Refactor of the code and added search Page
