+++ src/App.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 import "./App.css";
-import LoginForm from "./loginComponet/LoginForm";
+import "./common/index.css";
+import LoginForm from "./Pages/login/LoginForm";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import HomePage from "./HomePage";
-import SignUp from "./loginComponet/signUp";
+import HomePage from "./Pages/HomePage";
+import SignUp from "./Pages/signup/signUp";
+import HomeSearch from "./Pages/search/HomeSearch";
 
 class App extends React.Component {
   render() {
@@ -18,6 +20,9 @@ class App extends React.Component {
           </Route>
           <Route path="/signup">
             <SignUp />
+          </Route>
+          <Route path="/search">
+            <HomeSearch />
           </Route>
         </Switch>
       </Router>
