+++ src/Pages/search/homesearch.css
@@ -1,0 +1,7 @@
+.displayForm {
+  min-height: 100vh;
+  display: flex;
+  /* flex-direction: column;
+  align-items: center;
+  justify-content: center; */
+}
