+++ src/Pages/signup/signup.css
@@ -1,0 +1,3 @@
+.checkBox {
+  font-size: x-small;
+}
