+++ src/common/index.css
@@ -7,6 +7,7 @@
 }
 .border {
   box-shadow: 0px 1px 18px 0px #666;
+  width: 22rem;
 }
 
 .formTitle {
