--- src/loginComponet/signup.css
@@ -1,24 +1,0 @@
-.displayForm {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-.border {
-  box-shadow: 0px 1px 18px 0px #666;
-}
-
-.formTitle {
-  text-align: center;
-  padding: 30px;
-}
-.inputForm {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.checkBox {
-  font-size: x-small;
-}
